Review: extract prodcut definitions to model.ts
